Move add_lazy_dependency into queued.py
