+++ esp/esp/cache/queued.py
@@ -1,5 +1,7 @@
 """ General object with queuable actions. """
 
+import types
+
 def delay_method(func):
     def add_to_queue(self, *args, **kwargs):
         self._methods_queue.append( (func, args, kwargs) )
@@ -15,3 +17,14 @@ class WithDelayableMethods(object):
             func(self, *args, **kwargs)
         self._methods_queue = []
 
+def add_lazy_dependency(self, obj, operation):
+    """ If obj is a function (thunk), delay operation; otherwise execute immediately. """
+    # XXX: this is temporarily clunky because it's using delay_method in an unintended way.
+    # TODO: delete delay_method and do something better here.
+    if isinstance(obj, types.FunctionType):
+        @delay_method
+        def wrapped(self, obj):
+            return operation(self, obj())
+        wrapped(self, obj)
+    else:
+        operation(self, obj)
